requirements.txt
numpy 
scikit-learn
transformers
torch
fastapi
pydantic
tf-keras
bert-extractive-summarizer
uvicorn[standard]